huge refactoring (#5)
the goal of this refactoring is:

enable all possible currencies without a dedicated method for each
dramatically reduce the codebase size (~66%)
make it much simpler to add additional calls
this is accomplished by consolidating price calls inside one function call which queries the api for any provided currency code, silently failing if not found. calls are now defined in two dictionaries, itemDic and currencyFormatDic. To add a new item price or specific formatting for an already support currency simply add a definition to either dictionary.

in addition, several r/bitcoin specific moderation tools and utilities have been added to the utilities.py file. these should be generalized in a future refactoring to enable broader use of this bot.